% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boc.R
\name{boc_groups}
\alias{boc_groups}
\title{List available groups on the Bank of Canada Valet API}
\usage{
boc_groups()
}
\value{
tibble: name (character), label (character), description (character)
}
\description{
List available groups on the Bank of Canada Valet API
}
