% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boc.R
\name{boc_series}
\alias{boc_series}
\title{List available series in a Bank of Canada Valet group}
\usage{
boc_series(group)
}
\arguments{
\item{group}{Group name (e.g. "FX_RATES_DAILY", "FX_RATES_MONTHLY")}
}
\value{
tibble: name (character), label (character), description (character)
}
\description{
List available series in a Bank of Canada Valet group
}
